Centralize internal filters code
