+++ frontend/src/queries/nodes/DataNode/TestAccountFilters.tsx
@@ -1,6 +1,6 @@
 import { useActions, useValues } from 'kea'
+import { TestAccountFilterSwitch } from 'lib/components/TestAccountFiltersSwitch'
 import { filterTestAccountsDefaultsLogic } from 'scenes/settings/project/filterTestAccountDefaultsLogic'
-import { TestAccountFilterSwitch } from 'scenes/settings/project/TestAccountFiltersConfig'
 import { teamLogic } from 'scenes/teamLogic'
 
 import { DataNode, EventsQuery, HogQLQuery } from '~/queries/schema'
