+++ frontend/src/scenes/settings/project/TestAccountFiltersConfig.tsx
@@ -1,12 +1,10 @@
-import { IconGear } from '@posthog/icons'
-import { LemonButton, LemonSwitch, LemonSwitchProps, Link } from '@posthog/lemon-ui'
+import { LemonSwitch, Link } from '@posthog/lemon-ui'
 import { useActions, useValues } from 'kea'
 import { PropertyFilters } from 'lib/components/PropertyFilters/PropertyFilters'
 import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
 import { LemonBanner } from 'lib/lemon-ui/LemonBanner'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { teamLogic } from 'scenes/teamLogic'
-import { urls } from 'scenes/urls'
 
 import { groupsModel } from '~/models/groupsModel'
 import { AnyPropertyFilter } from '~/types'
@@ -111,38 +109,5 @@ export function ProjectAccountFiltersSetting(): JSX.Element {
             </div>
             <TestAccountFiltersConfig />
         </>
-    )
-}
-
-type TestAccountFilterProps = Partial<LemonSwitchProps> & {
-    checked: boolean
-    onChange: (checked: boolean) => void
-}
-
-export function TestAccountFilterSwitch({ checked, onChange, ...props }: TestAccountFilterProps): JSX.Element | null {
-    const { currentTeam } = useValues(teamLogic)
-    const hasFilters = (currentTeam?.test_account_filters || []).length > 0
-
-    return (
-        <LemonSwitch
-            id="test-account-filter"
-            bordered
-            disabledReason={!hasFilters ? "You haven't set any internal test filters" : null}
-            {...props}
-            checked={checked}
-            onChange={onChange}
-            label={
-                <div className="flex items-center">
-                    <span>Filter out internal and test users</span>
-                    <LemonButton
-                        icon={<IconGear />}
-                        to={urls.settings('project', 'internal-user-filtering')}
-                        size="small"
-                        noPadding
-                        className="ml-1"
-                    />
-                </div>
-            }
-        />
     )
 }
